fix retrofit mock, now it works
delete object retrofit services, create class from it, and inject
